requirements.txt
django==3.1.6
djangorestframework==3.12.2
dj-database-url==0.5.0
psycopg2-binary==2.8.6
pillow==8.1.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
pika==1.2.0